Finished Fuzzy K-Means
- Finished implementing fuzzy K-Means.  Dirtier implementation so as to
avoid Gaussian Distribution programming.  Added a basic header to the
page

- The Hierarchical Clustering requires many more iterations than the
other algorithms so either they are all slow or that one is really
slow… Not sure how to fix this but for rn just have a loop around the
interval that is a function of the number of nodes due to the algorithm
being O(n^2)?

- In the future I would like to go back and merge these algorithms so
as not have so much separate yet similar content.  Some of the
functions are identical but with different references (“_KMEANS” vs
“_FUZZY” vs “_HIGH”).  The reason for this is its easier to debug
because I’m keeping problems separate but it creates a lot of
superfluous code!!